Introduce BaseItem class
