+++ cobbler/cobbler_collections/collection.py
@@ -28,6 +28,7 @@ from cobbler.cexceptions import CX
 from cobbler.items import distro, image
 from cobbler.items import item as item_base
 from cobbler.items import menu, profile, repo, system
+from cobbler.items.abstract.base_item import BaseItem
 
 if TYPE_CHECKING:
     from cobbler.actions.sync import CobblerSync
@@ -35,7 +36,7 @@ if TYPE_CHECKING:
     from cobbler.cobbler_collections.manager import CollectionManager
 
 
-ITEM = TypeVar("ITEM", bound=item_base.Item)
+ITEM = TypeVar("ITEM", bound=BaseItem)
 FIND_KWARGS = Union[  # pylint: disable=invalid-name
     str, int, bool, Dict[Any, Any], List[Any]
 ]
@@ -164,7 +165,7 @@ class Collection(Generic[ITEM]):
         :returns: NotImplementedError
         """
 
-    def get(self, name: str) -> Optional[item_base.Item]:
+    def get(self, name: str) -> Optional[ITEM]:
         """
         Return object with name in the collection
 
@@ -367,7 +368,7 @@ class Collection(Generic[ITEM]):
                 raise ValueError("Unexepcted return value from find_items!")
             for item in items:
                 attr = getattr(item, "_" + dep_type[1])
-                if isinstance(attr, (str, item_base.Item)):
+                if isinstance(attr, (str, BaseItem)):
                     setattr(item, dep_type[1], newname)
                 elif isinstance(attr, list):
                     for i, attr_val in enumerate(attr):  # type: ignore
