+++ cobbler/items/item.py
@@ -100,27 +100,22 @@ V2.8.5:
 # SPDX-FileCopyrightText: Copyright 2006-2009, Red Hat, Inc and Others
 # SPDX-FileCopyrightText: Michael DeHaan <michael.dehaan AT gmail>
 
-import copy
-import enum
 import fnmatch
-import logging
 import pprint
-import re
 import uuid
-from abc import abstractmethod
+from abc import ABC
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple, Type, TypeVar, Union
-
-import yaml
 
 from cobbler import enums, utils
 from cobbler.cexceptions import CX
 from cobbler.decorator import InheritableDictProperty, InheritableProperty, LazyProperty
-from cobbler.items.abstract.item_cache import ItemCache
+from cobbler.items.abstract.base_item import BaseItem
 from cobbler.utils import input_converters
 
 if TYPE_CHECKING:
     from cobbler.api import CobblerAPI
-    from cobbler.cobbler_collections.collection import ITEM, ITEM_UNION
+    from cobbler.cobbler_collections.collection import ITEM_UNION
+    from cobbler.items.abstract.base_item import ITEM
     from cobbler.items.distro import Distro
     from cobbler.items.menu import Menu
     from cobbler.items.profile import Profile
@@ -128,11 +123,10 @@ if TYPE_CHECKING:
     from cobbler.settings import Settings
 
 
-RE_OBJECT_NAME = re.compile(r"[a-zA-Z0-9_\-.:]*$")
 T = TypeVar("T")
 
 
-class Item:
+class Item(BaseItem, ABC):
     """
     An Item is a serializable thing that can appear in a Collection
     """
@@ -252,27 +246,9 @@ class Item:
 
         raise TypeError(f"find cannot compare type: {type(from_obj)}")
 
-    @staticmethod
-    def __is_dict_key(name: str) -> bool:
-        """
-        Whether the attribute is part of the item's to_dict or not
-
-        :name: The attribute name.
-        """
-        return (
-            name[:1] == "_"
-            and "__" not in name
-            and name
-            not in {
-                "_last_cached_mtime",
-                "_cache",
-                "_supported_boot_loaders",
-                "_has_initialized",
-                "_inmemory",
-            }
-        )
-
-    def __init__(self, api: "CobblerAPI", is_subobject: bool = False, **kwargs: Any):
+    def __init__(
+        self, api: "CobblerAPI", *args: Any, is_subobject: bool = False, **kwargs: Any
+    ):
         """
         Constructor.  Requires a back reference to the CobblerAPI object.
 
@@ -303,17 +279,11 @@ class Item:
         :param api: The Cobbler API object which is used for resolving information.
         :param is_subobject: See above extensive description.
         """
-        # Prevent attempts to clear the to_dict cache before the object is initialized.
-        self._has_initialized = False
+        super().__init__(api, *args, **kwargs)
 
         self._parent = ""
         self._depth = 0
         self._children: List[str] = []
-        self._ctime = 0.0
-        self._mtime = 0.0
-        self._uid = uuid.uuid4().hex
-        self._name = ""
-        self._comment = ""
         self._kernel_options: Union[Dict[Any, Any], str] = {}
         self._kernel_options_post: Union[Dict[Any, Any], str] = {}
         self._autoinstall_meta: Union[Dict[Any, Any], str] = {}
@@ -321,13 +291,8 @@ class Item:
         self._boot_files: Union[Dict[Any, Any], str] = {}
         self._template_files: Dict[str, Any] = {}
         self._last_cached_mtime = 0
-        self._owners: Union[List[Any], str] = enums.VALUE_INHERITED
-        self._cache: ItemCache = ItemCache(api)
         self._is_subobject = is_subobject
         self._inmemory = True
-
-        self.logger = logging.getLogger()
-        self.api = api
 
         if len(kwargs) > 0:
             kwargs.update({"is_subobject": is_subobject})
@@ -338,27 +303,6 @@ class Item:
         if not self._has_initialized:
             self._has_initialized = True
 
-    def __eq__(self, other: Any) -> bool:
-        """
-        Comparison based on the uid for our items.
-
-        :param other: The other Item to compare.
-        :return: True if uid is equal, otherwise false.
-        """
-        if isinstance(other, Item):
-            return self._uid == other.uid
-        return False
-
-    def __hash__(self):
-        """
-        Hash table for Items.
-        Requires special handling if the uid value changes and the Item
-        is present in set, frozenset, and dict types.
-
-        :return: hash(uid).
-        """
-        return hash(self._uid)
-
     def __setattr__(self, name: str, value: Any):
         """
         Intercepting an attempt to assign a value to an attribute.
@@ -366,7 +310,7 @@ class Item:
         :name: The attribute name.
         :value: The attribute value.
         """
-        if Item.__is_dict_key(name) and self._has_initialized:
+        if Item._is_dict_key(name) and self._has_initialized:
             self.clean_cache(name)
         super().__setattr__(name, value)
 
@@ -501,130 +445,6 @@ class Item:
                 value.pop(key)  # type: ignore
 
         return value
-
-    @property
-    def uid(self) -> str:
-        """
-        The uid is the internal unique representation of a Cobbler object. It should never be used twice, even after an
-        object was deleted.
-
-        :getter: The uid for the item. Should be unique across a running Cobbler instance.
-        :setter: The new uid for the object. Should only be used by the Cobbler Item Factory.
-        """
-        return self._uid
-
-    @uid.setter
-    def uid(self, uid: str) -> None:
-        """
-        Setter for the uid of the item.
-
-        :param uid: The new uid.
-        """
-        if self._uid != uid and self.COLLECTION_TYPE != Item.COLLECTION_TYPE:
-            collection = self.api.get_items(self.COLLECTION_TYPE)
-            with collection.lock:
-                item = collection.get(self.name)
-                if item is not None and item.uid == self._uid:
-                    # Update uid index
-                    indx_dict = collection.indexes["uid"]
-                    del indx_dict[self._uid]
-                    indx_dict[uid] = self.name
-        self._uid = uid
-
-    @property
-    def ctime(self) -> float:
-        """
-        Property which represents the creation time of the object.
-
-        :getter: The float which can be passed to Python time stdlib.
-        :setter: Should only be used by the Cobbler Item Factory.
-        """
-        return self._ctime
-
-    @ctime.setter
-    def ctime(self, ctime: float) -> None:
-        """
-        Setter for the ctime property.
-
-        :param ctime: The time the object was created.
-        :raises TypeError: In case ``ctime`` was not of type float.
-        """
-        if not isinstance(ctime, float):  # type: ignore
-            raise TypeError("ctime needs to be of type float")
-        self._ctime = ctime
-
-    @property
-    def name(self) -> str:
-        """
-        Property which represents the objects name.
-
-        :getter: The name of the object.
-        :setter: Updating this has broad implications. Please try to use the ``rename()`` functionality from the
-                 corresponding collection.
-        """
-        return self._name
-
-    @name.setter
-    def name(self, name: str) -> None:
-        """
-        The objects name.
-
-        :param name: object name string
-        :raises TypeError: In case ``name`` was not of type str.
-        :raises ValueError: In case there were disallowed characters in the name.
-        """
-        if not isinstance(name, str):  # type: ignore
-            raise TypeError("name must of be type str")
-        if not RE_OBJECT_NAME.match(name):
-            raise ValueError(f"Invalid characters in name: '{name}'")
-        self._name = name
-
-    @LazyProperty
-    def comment(self) -> str:
-        """
-        For every object you are able to set a unique comment which will be persisted on the object.
-
-        :getter: The comment or an emtpy string.
-        :setter: The new comment for the item.
-        """
-        return self._comment
-
-    @comment.setter
-    def comment(self, comment: str) -> None:
-        """
-        Setter for the comment of the item.
-
-        :param comment: The new comment. If ``None`` the comment will be set to an emtpy string.
-        """
-        self._comment = comment
-
-    @InheritableProperty
-    def owners(self) -> List[Any]:
-        """
-        This is a feature which is related to the ownership module of Cobbler which gives only specific people access
-        to specific records. Otherwise this is just a cosmetic feature to allow assigning records to specific users.
-
-        .. warning:: This is never validated against a list of existing users. Thus you can lock yourself out of a
-                     record.
-
-        .. note:: This property can be set to ``<<inherit>>``.
-
-        :getter: Return the list of users which are currently assigned to the record.
-        :setter: The list of people which should be new owners. May lock you out if you are using the ownership
-                 authorization module.
-        """
-        return self._resolve("owners")
-
-    @owners.setter  # type: ignore[no-redef]
-    def owners(self, owners: Union[str, List[Any]]):
-        """
-        Setter for the ``owners`` property.
-
-        :param owners: The new list of owners. Will not be validated for existence.
-        """
-        if not isinstance(owners, (str, list)):  # type: ignore
-            raise TypeError("owners must be str or list!")
-        self._owners = input_converters.input_string_or_list(owners)
 
     @InheritableDictProperty
     def kernel_options(self) -> Dict[Any, Any]:
@@ -814,27 +634,6 @@ class Item:
         if not isinstance(depth, int):  # type: ignore
             raise TypeError("depth needs to be of type int")
         self._depth = depth
-
-    @property
-    def mtime(self) -> float:
-        """
-        Represents the last modification time of the object via the API. This is not updated automagically.
-
-        :getter: The float which can be fed into a Python time object.
-        :setter: The new time something was edited via the API.
-        """
-        return self._mtime
-
-    @mtime.setter
-    def mtime(self, mtime: float) -> None:
-        """
-        Setter for the modification time of the object.
-
-        :param mtime: The new modification time.
-        """
-        if not isinstance(mtime, float):  # type: ignore
-            raise TypeError("mtime needs to be of type float")
-        self._mtime = mtime
 
     @LazyProperty
     def parent(self) -> Optional[Union["System", "Profile", "Distro", "Menu"]]:
@@ -1104,142 +903,6 @@ class Item:
             return pprint.pformat(raw)
         return raw
 
-    def check_if_valid(self) -> None:
-        """
-        Raise exceptions if the object state is inconsistent.
-
-        :raises CX: In case the name of the item is not set.
-        """
-        if not self.name:
-            raise CX("Name is required")
-
-    @abstractmethod
-    def make_clone(self) -> "ITEM":  # type: ignore
-        """
-        Must be defined in any subclass
-        """
-        raise NotImplementedError("Must be implemented in a specific Item")
-
-    @classmethod
-    def _remove_depreacted_dict_keys(cls, dictionary: Dict[Any, Any]) -> None:
-        """
-        This method does remove keys which should not be deserialized and are only there for API compatibility in
-        :meth:`~cobbler.items.item.Item.to_dict`.
-
-        :param dictionary: The dict to update
-        """
-        if "ks_meta" in dictionary:
-            dictionary.pop("ks_meta")
-        if "kickstart" in dictionary:
-            dictionary.pop("kickstart")
-        if "children" in dictionary:
-            dictionary.pop("children")
-
-    def from_dict(self, dictionary: Dict[Any, Any]) -> None:
-        """
-        Modify this object to take on values in ``dictionary``.
-
-        :param dictionary: This should contain all values which should be updated.
-        :raises AttributeError: In case during the process of setting a value for an attribute an error occurred.
-        :raises KeyError: In case there were keys which could not be set in the item dictionary.
-        """
-        self._remove_depreacted_dict_keys(dictionary)
-        if len(dictionary) == 0:
-            return
-        old_has_initialized = self._has_initialized
-        self._has_initialized = False
-        result = copy.deepcopy(dictionary)
-        for key in dictionary:
-            lowered_key = key.lower()
-            # The following also works for child classes because self is a child class at this point and not only an
-            # Item.
-            if hasattr(self, "_" + lowered_key):
-                try:
-                    setattr(self, lowered_key, dictionary[key])
-                except AttributeError as error:
-                    raise AttributeError(
-                        f'Attribute "{lowered_key}" could not be set!'
-                    ) from error
-                result.pop(key)
-        self._has_initialized = old_has_initialized
-        self.clean_cache()
-        if len(result) > 0:
-            raise KeyError(
-                f"The following keys supplied could not be set: {result.keys()}"
-            )
-
-    def to_dict(self, resolved: bool = False) -> Dict[Any, Any]:
-        """
-        This converts everything in this object to a dictionary.
-
-        :param resolved: If this is True, Cobbler will resolve the values to its final form, rather than give you the
-                     objects raw value.
-        :return: A dictionary with all values present in this object.
-        """
-        if not self.inmemory:
-            self.deserialize()
-        cached_result = self.cache.get_dict_cache(resolved)
-        if cached_result is not None:
-            return cached_result
-
-        value: Dict[str, Any] = {}
-        for key, key_value in self.__dict__.items():
-            if self.__is_dict_key(key):
-                new_key = key[1:].lower()
-                if isinstance(key_value, enum.Enum):
-                    if resolved:
-                        value[new_key] = getattr(self, new_key).value
-                    else:
-                        value[new_key] = key_value.value
-                elif new_key == "interfaces":
-                    # This is the special interfaces dict. Lets fix it before it gets to the normal process.
-                    serialized_interfaces = {}
-                    interfaces = key_value
-                    for interface_key in interfaces:
-                        serialized_interfaces[interface_key] = interfaces[
-                            interface_key
-                        ].to_dict(resolved)
-                    value[new_key] = serialized_interfaces
-                elif isinstance(key_value, list):
-                    value[new_key] = copy.deepcopy(key_value)  # type: ignore
-                elif isinstance(key_value, dict):
-                    if resolved:
-                        value[new_key] = getattr(self, new_key)
-                    else:
-                        value[new_key] = copy.deepcopy(key_value)  # type: ignore
-                elif (
-                    isinstance(key_value, str)
-                    and key_value == enums.VALUE_INHERITED
-                    and resolved
-                ):
-                    value[new_key] = getattr(self, key[1:])
-                else:
-                    value[new_key] = key_value
-        if "autoinstall" in value:
-            value.update({"kickstart": value["autoinstall"]})  # type: ignore
-        if "autoinstall_meta" in value:
-            value.update({"ks_meta": value["autoinstall_meta"]})
-        self.cache.set_dict_cache(value, resolved)
-        return value
-
-    def serialize(self) -> Dict[str, Any]:
-        """
-        This method is a proxy for :meth:`~cobbler.items.item.Item.to_dict` and contains additional logic for
-        serialization to a persistent location.
-
-        :return: The dictionary with the information for serialization.
-        """
-        keys_to_drop = [
-            "kickstart",
-            "ks_meta",
-            "remote_grub_kernel",
-            "remote_grub_initrd",
-        ]
-        result = self.to_dict()
-        for key in keys_to_drop:
-            result.pop(key, "")
-        return result
-
     def deserialize(self) -> None:
         """
         Deserializes the object itself and, if necessary, recursively all the objects it depends on.
@@ -1271,14 +934,14 @@ class Item:
                                 deserialize_ancestor(ancestor_item_type, ancestor_name)
         self.from_dict(item_dict)
 
-    def grab_tree(self) -> List[Union["Item", "Settings"]]:
+    def grab_tree(self) -> List[Union["ITEM", "Settings"]]:
         """
         Climb the tree and get every node.
 
         :return: The list of items with all parents from that object upwards the tree. Contains at least the item
                  itself and the settings of Cobbler.
         """
-        results: List[Union["Item", "Settings"]] = [self]
+        results: List[Union["ITEM", "Settings"]] = [self]
         parent = self.logical_parent
         while parent is not None:
             results.append(parent)
@@ -1291,23 +954,10 @@ class Item:
         )
         return results
 
-    @property
-    def cache(self) -> ItemCache:
-        """
-        Gettinging the ItemCache oject.
-
-        .. note:: This is a read only property.
-
-        :getter: This is the ItemCache oject.
-        """
-        return self._cache
-
     def _clean_dict_cache(self, name: Optional[str]):
         """
         Clearing the Item dict cache.
 
-        :param obj: The object whose modification invalidates the dict cache.
-                    Can be Item, Settings or SIGNATURE_CACHE.
         :param name: The name of Item attribute or None.
         """
         if not self.api.settings().cache_enabled:
@@ -1326,34 +976,3 @@ class Item:
 
         # Invalidating the cache of the object itself.
         self.cache.clean_dict_cache()
-
-    def clean_cache(self, name: Optional[str] = None):
-        """
-        Clearing the Item cache.
-
-        :param obj: The object whose modification invalidates the dict cache.
-                    Can be Item, Settings or SIGNATURE_CACHE.
-        :param name: The name of Item attribute or None.
-        """
-        if self._inmemory:
-            self._clean_dict_cache(name)
-
-    @property
-    def inmemory(self) -> bool:
-        r"""
-        If set to ``false``, only the Item name is in memory. The rest of the Item's properties can be retrieved
-        either on demand or as a result of the ``load_items`` background task.
-
-        :getter: The inmemory for the item.
-        :setter: The new inmemory value for the object. Should only be used by the Cobbler serializers.
-        """
-        return self._inmemory
-
-    @inmemory.setter
-    def inmemory(self, inmemory: bool):
-        """
-        Setter for the inmemory of the item.
-
-        :param inmemory: The new inmemory value.
-        """
-        self._inmemory = inmemory
