+++ cobbler/validate.py
@@ -15,7 +15,7 @@ from uuid import UUID
 import netaddr
 
 from cobbler import enums, utils
-from cobbler.items import item
+from cobbler.items.abstract import base_item
 from cobbler.utils import input_converters, signatures
 
 if TYPE_CHECKING:
@@ -628,4 +628,4 @@ def validate_obj_name(object_name: str) -> bool:
     """
     if not isinstance(object_name, str):  # type: ignore
         return False
-    return bool(re.fullmatch(item.RE_OBJECT_NAME, object_name))
+    return bool(re.fullmatch(base_item.RE_OBJECT_NAME, object_name))
